move Py3 to own file
